# docs/openapi3.yaml
openapi: 3.0.0
info:
  description: Production-ready backend Endge-сервиса с эталонными
    RedPanda-обёртками и строгими архитектурными границами.
  title: Endge Service Backend API
  contact: {}
  version: 1.0.20
paths:
...
  "/api/v1/projects/{project_identity}/data-views":
    get:
      security:
        - BearerAuth: []
      description: Возвращает активные DataView проекта с optional фильтрами папки и
        Query. DataView, связанные с soft-deleted Query, не возвращаются.
      tags:
        - data-views
      summary: Список DataView
      parameters:
        - example: demo-project
          description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
        - example: root-data-views
          description: Folder identity
          name: folder_identity
          in: query
          schema:
            type: string
        - example: users-list
          description: Query identity
          name: query_identity
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_http_v1_data_view.DataViewsListRespons\
                  e"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
    post:
      security:
        - BearerAuth: []
      description: Создает DataView в папке проекта и связывает его с активной Query
        этого же проекта. Source и schema хранятся как authoring configuration и
        не компилируются и не выполняются.
      tags:
        - data-views
      summary: Создать DataView
      parameters:
        - example: demo-project
          description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_api_http_v1_data_view.CreateDataViewReques\
                t"
        description: Параметры DataView
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_http_v1_data_view.DataViewResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
  "/api/v1/projects/{project_identity}/data-views/{data_view_identity}":
    get:
      security:
        - BearerAuth: []
      description: Возвращает активный DataView по identity в пределах проекта вместе
        с identity папки и связанной Query.
      tags:
        - data-views
      summary: Получить DataView
      parameters:
        - example: demo-project
          description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
        - example: users-table-view
          description: DataView identity
          name: data_view_identity
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_http_v1_data_view.DataViewResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
    delete:
      security:
        - BearerAuth: []
      description: Выполняет soft-delete DataView по identity в пределах проекта.
      tags:
        - data-views
      summary: Удалить DataView
      parameters:
        - description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
        - description: DataView identity
          name: data_view_identity
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
    patch:
      security:
        - BearerAuth: []
      description: Полностью заменяет editable payload DataView, включая папку,
        связанную Query и authoring configuration. Поля id, identity, createdAt
        и deletedAt сохраняются.
      tags:
        - data-views
      summary: Обновить DataView
      parameters:
        - example: demo-project
          description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
        - example: users-table-view
          description: DataView identity
          name: data_view_identity
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_api_http_v1_data_view.UpdateDataViewReques\
                t"
        description: Параметры обновления DataView
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_http_v1_data_view.DataViewResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
  "/api/v1/projects/{project_identity}/data-views/{data_view_identity}/hard":
    delete:
      security:
        - BearerAuth: []
      description: Выполняет hard-delete soft-deleted DataView по identity в пределах
        проекта.
      tags:
        - data-views
      summary: Физически удалить DataView
      parameters:
        - description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
        - description: DataView identity
          name: data_view_identity
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
  "/api/v1/projects/{project_identity}/data-views/{data_view_identity}/restore":
    post:
      security:
        - BearerAuth: []
      description: Восстанавливает soft-deleted DataView по identity в пределах проекта.
      tags:
        - data-views
      summary: Восстановить DataView
      parameters:
        - description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
        - description: DataView identity
          name: data_view_identity
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
...
servers:
  - url: /
components:
  securitySchemes:
    BearerAuth:
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    internal_api_http_v1_data_view.DataViewResponse:
      type: object
      properties:
        active:
          type: boolean
        createdAt:
          type: string
        deletedAt:
          type: string
        description:
          type: string
        displayName:
          type: string
        folderIdentity:
          type: string
        id:
          type: string
        identity:
          type: string
        inputSchema:
          type: object
          additionalProperties: {}
        meta:
          type: object
          additionalProperties: {}
        outputSchema:
          type: object
          additionalProperties: {}
        projectIdentity:
          type: string
        queryIdentity:
          type: string
        source:
          type: object
          additionalProperties: {}
        updatedAt:
          type: string
        viewType:
          type: string
...